Integrate PHP to Follows HTML (#21)
* Integrate PHP to Follows HTML

Attempt at Making calls to the database with the follows page. It is non-functional at the moment, so do not merge this branch with main.

* Follows page now PHP

Follows page is now PHP with correct number of calls to the database (don't merge, not functioning)

* connect follows page to database 

Follows page can now connect to the database using new $conn variable.